feat: lazy loading huge main page components (#449)
* feat: lazy loading huge main page components

* fix: package-lock

* fix: lint

* fix: typecheck

* fix: issues

* fix: fetch contributors from server

* fix: client and server caches

* fix: conflicts

* fix: issues
