+++ src/components/ExpandableContributorList/index.ts
@@ -1,1 +1,2 @@
 export * from './ExpandableContributorList';
+export {LazyExpandableContributorsList} from './LazyExpandableContributorsList';
