+++ src/components/UISamples/TablePreview/TablePreview.scss
@@ -16,4 +16,13 @@ $block: '.#{variables.$ns}table-preview';
         overflow-x: auto;
         white-space: nowrap;
     }
+
+    &__loader {
+        width: 100%;
+        height: 100%;
+
+        display: flex;
+        justify-content: center;
+        align-items: center;
+    }
 }
