+++ src/components/UISamples/UISamplesLoader/UISamplesLoader.scss
@@ -1,0 +1,16 @@
+@use '../../../variables';
+@use '~@gravity-ui/page-constructor/styles/variables.scss' as pcVariables;
+
+$block: '.#{variables.$ns}uisamples-loader';
+
+#{$block} {
+    width: 100%;
+    height: 800px;
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    border: 1px solid var(--g-color-line-generic);
+    border-radius: 16px;
+}
