+++ src/components/UISamples/index.ts
@@ -1,13 +1,20 @@
 export {ApartmentCardPreview} from './ApartmentCardPreview/ApartmentCardPreview';
+export {LazyApartmentCardPreview} from './ApartmentCardPreview/LazyApartmentCardPreview';
 export {CardsPreview} from './CardsPreview/CardsPreview';
 export {DashboardPreview2} from './DashboardPreview2/DashboardPreview2';
+export {LazyDashboardPreview2} from './DashboardPreview2/LazyDashboardPreview2';
 export {DashboardPreview} from './DashboardsPreview/DashboardPreview';
 export {FormPreview} from './FormPreview/FormPreview';
 export {KubernetesPreview} from './KubernetesPreview/KubernetesPreview';
+export {LazyKubernetesPreview} from './KubernetesPreview/LazyKubernetesPreview';
 export {MailPreview} from './MailPreview/MailPreview';
+export {LazyMailPreview} from './MailPreview/LazyMailPreview';
 export {OsnPreview} from './OsnPreview/OsnPreview';
+export {LazyOsnPreview} from './OsnPreview/LazyOsnPreview';
 export {PreviewLayout} from './PreviewLayout/PreviewLayout';
 export {PreviewWrapper} from './PreviewWrapper/PreviewWrapper';
 export {TablePreview} from './TablePreview/TablePreview';
+export {LazyTablePreview} from './TablePreview/LazyTablePreview';
 export {TasksPreview} from './TasksPreview/TasksPreview';
+export {LazyTasksPreview} from './TasksPreview/LazyTasksPreview';
 export * from './constants';
