+++ src/content/landing.ts
@@ -1,7 +1,7 @@
 import {BlockType} from '@gravity-ui/page-constructor';
 import {TFunction} from 'next-i18next';
 
-import type {Contributor, LibWithMetadata} from '../api';
+import type {LibWithMetadata} from '../api';
 import companiesDesktopAsset from '../assets/companies-desktop.svg';
 import companiesMobileAsset from '../assets/companies-mobile.svg';
 import companiesTabletAsset from '../assets/companies-tablet.svg';
@@ -20,12 +20,10 @@ import {CustomPageContent} from './types';
 export const getLanding = ({
     t,
     libs,
-    contributors,
     backgroundImageSrc,
 }: {
     t: TFunction;
     libs: LibWithMetadata[];
-    contributors: Contributor[];
     backgroundImageSrc: string;
 }): CustomPageContent => ({
     background: {
@@ -185,7 +183,6 @@ export const getLanding = ({
                 title: t('home:contributors_actions_telegram'),
                 href: 'https://t.me/gravity_ui',
             },
-            contributors,
         },
         {
             type: BlockType.CompaniesBlock,
