+++ src/pages/components/[libId]/[componentId].tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import {Section} from 'src/components/NavigationLayout/types';
 
 import i18nextConfig from '../../../../next-i18next.config';
-import {Api, type LibWithFullData} from '../../../api';
+import {type LibWithFullData, ServerApi} from '../../../api';
 import {Component} from '../../../components/Component/Component';
 import {ComponentsLayout} from '../../../components/ComponentsLayout/ComponentsLayout';
 import {Layout} from '../../../components/Layout/Layout';
@@ -34,9 +34,9 @@ export const getServerSideProps: GetServerSideProps = async (ctx) => {
 
     const libId = ctx.params?.libId as string;
 
-    const libPromise = Api.instance.fetchLibByIdWithCache(libId);
+    const libPromise = ServerApi.instance.fetchLibByIdWithCache(libId);
     const i18nPropsPromise = getI18nProps(ctx, ['component', 'libraries-info', 'component-meta']);
-    const readmePromise = Api.instance.fetchComponentReadmeWithCache({
+    const readmePromise = ServerApi.instance.fetchComponentReadmeWithCache({
         readmeUrl: component.content.readmeUrl,
         componentId: component.id,
         libId,
