+++ src/pages/index.tsx
@@ -1,30 +1,28 @@
 import {GetServerSideProps} from 'next';
 import Head from 'next/head';
 
-import type {Contributor, LibWithMetadata} from '../api';
-import {Api} from '../api';
+import type {LibWithMetadata} from '../api';
+import {ServerApi} from '../api';
 import backgroundAsset from '../assets/background.jpg';
 import {Landing} from '../components/Landing/Landing';
 import {Layout} from '../components/Layout/Layout';
 import {getI18nProps} from '../utils/i18next';
 
 export const getServerSideProps: GetServerSideProps = async (ctx) => {
-    const [contributors, libs, i18nProps] = await Promise.all([
-        Api.instance.fetchAllContributorsWithCache(),
-        Api.instance.fetchLandingLibs(),
+    const [libs, i18nProps] = await Promise.all([
+        ServerApi.instance.fetchLandingLibs(),
         getI18nProps(ctx, ['home', 'libraries-info']),
     ]);
 
     return {
         props: {
-            contributors,
             libs,
             ...i18nProps,
         },
     };
 };
 
-const Home = ({libs, contributors}: {libs: LibWithMetadata[]; contributors: Contributor[]}) => {
+const Home = ({libs}: {libs: LibWithMetadata[]}) => {
     return (
         <>
             <Head>
@@ -32,11 +30,7 @@ const Home = ({libs, contributors}: {libs: LibWithMetadata[]; contributors: Cont
                 <link rel="preload" as="image" href={backgroundAsset.src} type="image/jpeg" />
             </Head>
             <Layout isPageConstructor>
-                <Landing
-                    libs={libs}
-                    contributors={contributors}
-                    backgroundImageSrc={backgroundAsset.src}
-                />
+                <Landing libs={libs} backgroundImageSrc={backgroundAsset.src} />
             </Layout>
         </>
     );
