+++ src/pages/rtl.tsx
@@ -4,35 +4,27 @@ import Head from 'next/head';
 import React from 'react';
 
 import nextI18nextConfig from '../../next-i18next.config';
-import {Api, Contributor, LibWithMetadata} from '../api';
+import {LibWithMetadata, ServerApi} from '../api';
 import backgroundAsset from '../assets/background.jpg';
 import {Landing} from '../components/Landing/Landing';
 import {Layout} from '../components/Layout/Layout';
 import {getI18nProps} from '../utils/i18next';
 
 export const getServerSideProps: GetServerSideProps = async (ctx) => {
-    const [contributors, libs, i18nProps] = await Promise.all([
-        Api.instance.fetchAllContributorsWithCache(),
-        Api.instance.fetchLandingLibs(),
+    const [libs, i18nProps] = await Promise.all([
+        ServerApi.instance.fetchLandingLibs(),
         getI18nProps(ctx, ['home', 'libraries-info']),
     ]);
 
     return {
         props: {
-            contributors,
             libs,
             ...i18nProps,
         },
     };
 };
 
-export const RTLPage = ({
-    libs,
-    contributors,
-}: {
-    libs: LibWithMetadata[];
-    contributors: Contributor[];
-}) => {
+export const RTLPage = ({libs}: {libs: LibWithMetadata[]}) => {
     const {i18n} = useTranslation();
     i18n.changeLanguage(nextI18nextConfig.i18n.defaultLocale);
 
@@ -43,11 +35,7 @@ export const RTLPage = ({
                 <link rel="preload" as="image" href={backgroundAsset.src} type="image/jpeg" />
             </Head>
             <Layout isPageConstructor isRtl hideLocalePicker>
-                <Landing
-                    libs={libs}
-                    contributors={contributors}
-                    backgroundImageSrc={backgroundAsset.src}
-                />
+                <Landing libs={libs} backgroundImageSrc={backgroundAsset.src} />
             </Layout>
         </>
     );
